\name{BayesF2D-package}
\alias{BayesF2D-package}
\alias{BayesF2D}
\docType{package}
\title{
\packageTitle{BayesF2D}
}
\description{
\packageDescription{BayesF2D}
}
%\details{
%The DESCRIPTION file:
%\packageDESCRIPTION{BayesF2D}
%\packageIndices{BayesF2D}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
%}
\author{
\packageAuthor{BayesF2D}

Maintainer: \packageMaintainer{BayesF2D}
}
\references{
Thierry Chekouo et al (2020), \emph{A Bayesian 2D Functional Linear Model for Gray Level Co-occurrence Matrices in Texture Analysis of LowerGrade Gliomas}, \emph{submitted}
}

\keyword{ package }
\seealso{
\code{\link{BayesF2D}}
}
%\examples{
%~~ simple examples of the most important functions ~~
%}
