\docType{data}
\name{nonsymCor}
\alias{nonsymCor}
\title{
Empirical correlation matrices between non-symmetric GLCMs of the 4 imaging sequences.
}
\format{A list of 1 data frame:
\describe{
  \item{nonsymCor}{A list of correlation matrix for each imaging sequence
}
}
}
\usage{
nonsymCor
}
\description{
Empirical correlation matrices between GLCMs of the 4 imaging sequences. These values were used to simulate non symmetric GLCM matrices
}
\examples{
library(BayesF2D)
data("nonsymCor");
}



