\docType{data}
\name{symCor}
\alias{symCor}
\title{
Empirical correlation matrices between symmetric GLCMs of the 4 imaging sequences.
}
\format{A list of 1 data frame:
\describe{
  \item{symCor}{A list of correlation matrix for each imaging sequence
}
}
}
\usage{
symCor
}
\description{
Empirical correlation matrices between GLCMs of the 4 imaging sequences. These values were used to simulate symmetric GLCM matrices
}
\examples{
library(BayesF2D)
data("symCor");
}



